Moved histogram to dask.array.core
Also:

- correct dtype with/without weights
- better naming
- tests refactored and splitted
- cleanup unused references from wrap module
