+++ dask/array/__init__.py
@@ -3,8 +3,8 @@ from __future__ import absolute_import, division, print_function
 from ..utils import ignoring
 from .core import (Array, stack, concatenate, take, tensordot, transpose,
         from_array, choose, where, coarsen, insert, broadcast_to, fromfunction,
-        compute, unique, store, squeeze, topk, bincount, map_blocks, atop,
-        to_hdf5)
+        compute, unique, store, squeeze, topk, bincount, histogram, map_blocks,
+        atop, to_hdf5)
 from .core import (logaddexp, logaddexp2, conj, exp, log, log2, log10, log1p,
         expm1, sqrt, square, sin, cos, tan, arcsin, arccos, arctan, arctan2,
         hypot, sinh, cosh, tanh, arcsinh, arccosh, arctanh, deg2rad, rad2deg,
@@ -22,7 +22,7 @@ from .percentile import percentile
 with ignoring(ImportError):
     from .reductions import nanprod
 from . import random, linalg, ghost, learn
-from .wrap import ones, zeros, empty, full, histogram
+from .wrap import ones, zeros, empty, full
 from .rechunk import rechunk
 from ..context import set_options
 from .optimization import optimize
