+++ dask/array/tests/test_array_core.py
@@ -1007,6 +1007,65 @@ def test_bincount_raises_informative_error_on_missing_minlength_kwarg():
         assert False
 
 
+def test_histogram():
+    # Test for normal, flattened input
+    n = 100
+    v = da.random.random(n, chunks=10)
+    bins = np.arange(0, 1.01, 0.01)
+    (a1, b1) = da.histogram(v, bins=bins)
+    (a2, b2) = np.histogram(v, bins=bins)
+    
+    # Check if the sum of the bins equals the number of samples
+    assert a2.sum(axis=0) == n
+    assert a1.sum(axis=0) == n
+    eq(a1, a2)
+
+
+def test_histogram_alternative_bins_range():
+    v = da.random.random(100, chunks=10)
+    bins = np.arange(0, 1.01, 0.01)
+    # Other input
+    (a1, b1) = da.histogram(v, bins=10, range=(0, 1))
+    (a2, b2) = np.histogram(v, bins=10, range=(0, 1))
+    eq(a1, a2)
+    eq(b1, b2)
+
+
+def test_histogram_return_type():
+    v = da.random.random(100, chunks=10)
+    bins = np.arange(0, 1.01, 0.01)
+    # Check if return type is same as hist
+    bins = np.arange(0, 11, 1, dtype='i4')
+    (a1, b1) = da.histogram(v * 10, bins=bins)
+    (a2, b2) = np.histogram(v * 10, bins=bins)
+    eq(a1, a2)
+    assert a1.dtype == a2.dtype
+
+
+def test_histogram_extra_args_and_shapes():
+    # Check for extra args and shapes
+    bins = np.arange(0, 1.01, 0.01)
+    v = da.random.random(100, chunks=10)
+    data = [(v, bins, da.ones(100, chunks=v.chunks) * 5),
+            (da.random.random((50, 50), chunks=10), bins, da.ones((50, 50), chunks=10) * 5)]
+    
+    for v, bins, w in data:
+        # density
+        eq(da.histogram(v, bins=bins, normed=True)[0],
+           np.histogram(v, bins=bins, normed=True)[0])
+        
+        # normed
+        eq(da.histogram(v, bins=bins, density=True)[0],
+           np.histogram(v, bins=bins, density=True)[0])
+        
+        # weights
+        eq(da.histogram(v, bins=bins, weights=w)[0],
+           np.histogram(v, bins=bins, weights=w)[0])
+        
+        eq(da.histogram(v, bins=bins, weights=w, density=True)[0],
+           da.histogram(v, bins=bins, weights=w, density=True)[0])
+
+
 def test_concatenate3():
     x = np.array([1, 2])
     assert eq(concatenate3([x, x, x]),
