+++ dask/array/wrap.py
@@ -6,8 +6,7 @@ from functools import partial
 from toolz import curry
 import numpy as np
 
-from ..core import flatten
-from .core import Array, normalize_chunks, tokens, from_array
+from .core import Array, normalize_chunks, tokens
 from .numpy_compat import full
 
 names = ('wrapped_%d' % i for i in count(1))
@@ -113,75 +112,3 @@ ones = w(np.ones, dtype='f8')
 zeros = w(np.zeros, dtype='f8')
 empty = w(np.empty, dtype='f8')
 full = w(full)
-
-def histogram(a, bins=None, range=None, normed=False, weights=None, density=None):
-    """
-    Blocked variant of numpy.histogram.
-    
-    Follows the signature of numpy.histogram exactly with the following
-    exceptions:
-    
-    - either the ``bins`` or ``range`` argument is required as computing 
-      ``min`` and ``max`` over blocked arrays is an expensive operation 
-      that must be performed explicitly.
-    
-    - ``weights`` must be a dask.array.Array with the same block structure
-       as ``a``.
-    
-    Original signature follows below.
-    """ + np.histogram.__doc__
-    if bins is None or (range is None and bins is None):
-        raise ValueError('dask.array.histogram requires either bins '
-                         'or bins and range to be defined.')
-    
-    if weights is not None and weights.chunks != a.chunks:
-        raise ValueError('Input array and weights must have the same'
-                         'chunked structure')
-    
-    if not np.iterable(bins):
-        mn, mx = range
-        if mn == mx:
-            mn -= 0.5
-            mx += 0.5
-        
-        bins = np.linspace(mn, mx, bins + 1, endpoint=True)
-    
-    nchunks = len(list(flatten(a._keys())))
-    chunks = ((1,) * nchunks, (len(bins) - 1,))
-    
-    name1 = 'histogram-' + next(tokens)
-    
-    
-    # Map the histogram to all bins
-    def block_hist(x, weights=None):
-        return np.histogram(x, bins, weights=weights)[0][np.newaxis]
-    
-    if weights is None:
-        dsk = dict(((name1, i, 0), (block_hist, k))
-                    for i, k in enumerate(flatten(a._keys())))
-    else:
-        a_keys = flatten(a._keys())
-        w_keys = flatten(weights._keys())
-        dsk = dict(((name1, i, 0), (block_hist, k, w)) 
-                    for i, (k, w) in enumerate(zip(a_keys, w_keys)))
-        dsk.update(weights.dask)
-
-    dsk.update(a.dask)
-    
-    mapped = Array(dsk, name1, chunks, dtype=bins.dtype)
-    n = mapped.sum(axis=0)
-    
-    # We need to replicate normed and density options from numpy
-    if density is not None:
-        if density:
-            db = from_array(np.diff(bins).astype(float), chunks=n.chunks)
-            return n/db/n.sum(), bins
-        else:
-            return n, bins
-    else:
-        # deprecated, will be removed from Numpy 2.0
-        if normed:
-            db = from_array(np.diff(bins).astype(float), chunks=n.chunks)
-            return n/(n*db).sum(), bins
-        else:
-            return n, bins
